refactor(agents): move useLabels from lib/hooks to lib/agents/hooks
