+++ web/src/lib/hooks.ts
@@ -27,7 +27,7 @@ import { parseLlmDescriptor } from "@/lib/languageModels/utils";
 import { ChatSession } from "@/app/app/interfaces";
 import { Credential } from "./connectors/credentials";
 import { SettingsContext } from "@/providers/SettingsProvider";
-import { MinimalAgent, AgentLabel } from "@/lib/agents/types";
+import { MinimalAgent } from "@/lib/agents/types";
 import { DefaultModel, LLMProviderDescriptor } from "@/interfaces/llm";
 import { isAnthropic } from "@/lib/languageModels/svc";
 import { getSourceMetadataForSources } from "./sources";
@@ -244,85 +244,6 @@ export const useFederatedConnectors = () => {
   return {
     ...swrResponse,
     refreshFederatedConnectors: () => mutate(url),
-  };
-};
-
-export const useLabels = () => {
-  const { mutate } = useSWRConfig();
-  const { data: labels, error } = useSWR<AgentLabel[]>(
-    SWR_KEYS.personaLabels,
-    errorHandlingFetcher
-  );
-
-  const refreshLabels = async () => {
-    return mutate(SWR_KEYS.personaLabels);
-  };
-
-  const createLabel = async (name: string): Promise<AgentLabel | null> => {
-    const response = await fetch(SWR_KEYS.personaLabels, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ name }),
-    });
-
-    if (!response.ok) {
-      return null;
-    }
-
-    const newLabel: AgentLabel = await response.json();
-    mutate(
-      SWR_KEYS.personaLabels,
-      (currentLabels: AgentLabel[] | undefined) => [
-        ...(currentLabels || []),
-        newLabel,
-      ],
-      false
-    );
-    return newLabel;
-  };
-
-  const updateLabel = async (id: number, name: string) => {
-    const response = await fetch(`/api/admin/persona/label/${id}`, {
-      method: "PATCH",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ label_name: name }),
-    });
-
-    if (response.ok) {
-      mutate(
-        SWR_KEYS.personaLabels,
-        labels?.map((label) => (label.id === id ? { ...label, name } : label)),
-        false
-      );
-    }
-
-    return response;
-  };
-
-  const deleteLabel = async (id: number) => {
-    const response = await fetch(`/api/admin/persona/label/${id}`, {
-      method: "DELETE",
-      headers: { "Content-Type": "application/json" },
-    });
-
-    if (response.ok) {
-      mutate(
-        SWR_KEYS.personaLabels,
-        labels?.filter((label) => label.id !== id),
-        false
-      );
-    }
-
-    return response;
-  };
-
-  return {
-    labels,
-    error,
-    refreshLabels,
-    createLabel,
-    updateLabel,
-    deleteLabel,
   };
 };
