+++ web/src/sections/modals/ShareAgentModal.tsx
@@ -22,10 +22,9 @@ import useShareableGroups from "@/hooks/useShareableGroups";
 import { useModal } from "@/refresh-components/contexts/ModalContext";
 import { useUser } from "@/providers/UserProvider";
 import { Formik, useFormikContext } from "formik";
-import { useAgent } from "@/lib/agents/hooks";
+import { useAgent, useLabels } from "@/lib/agents/hooks";
 import { Button, Card, Divider, MessageCard, Text } from "@opal/components";
 import { Disabled } from "@opal/core";
-import { useLabels } from "@/lib/hooks";
 import { AgentLabel } from "@/lib/agents/types";
 import { FetchError } from "@/lib/fetcher";
